% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_function.R
\name{corr}
\alias{corr}
\title{corr}
\usage{
corr(sim)
}
\arguments{
\item{sim}{The dataframe or list produced by the 'runtsir' function.}
}
\description{
Plot the correlation of the true data against the fitted resimulated data.
}
