% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/derivative_function.R
\name{derivative}
\alias{derivative}
\title{derivative Function}
\usage{
derivative(X, Y)
}
\description{
This function computes an 8 point derivative
Differentiate Y with respect to X
}

