% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/epitimes_function.R
\name{epitimes}
\alias{epitimes}
\title{epitimes Function}
\usage{
epitimes(data, threshold)
}
\arguments{
\item{data}{dataframe with the cases vector.}

\item{threshold}{the required number of cases to call an outbreak.}
}
\description{
epitimes Function
}

