% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/jagsfilter_function.R
\name{jagsfilter}
\alias{jagsfilter}
\title{jagsfilter Function
used internally to filter jags results to give just the stuff well use}
\usage{
jagsfilter(mcmcresults)
}
\arguments{
\item{mcmcresults}{is the input from the jags model}
}
\description{
jagsfilter Function
used internally to filter jags results to give just the stuff well use
}

