% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/loadUKdata_function.R
\name{loadUKdata}
\alias{loadUKdata}
\title{loadUKdata function}
\usage{
loadUKdata(x)
}
\arguments{
\item{x,}{town name. Options are
"Bedwellty", "Birmingham", "Bradford", "Bristol", "Cardiff","Consett","Dalton.in.Furness" ,
"Halesworth", "Hastings", "Hull", "Leeds", "Lees", "Liverpool", "London",
"Manchester","Mold", "Northwich", "Nottingham", "Oswestry", "Sheffield"}
}
\description{
function to load and interpolate the births and population of 20 UK cities.
puts the data into biweeks.
}

