% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/loadvestmannaeyjar_function.R
\name{loadvestmannaeyjar}
\alias{loadvestmannaeyjar}
\title{loadvestmannaeyjar function}
\usage{
loadvestmannaeyjar()
}
\description{
function to load data from vestmannaeyjar, Iceland.
the period here is 24 weeks, therefore IP should be (52/24)
}

