% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/loadvopnafjardarhreppur_function.R
\name{loadvopnafjardarhreppur}
\alias{loadvopnafjardarhreppur}
\title{loadvestmannaeyjar function}
\usage{
loadvopnafjardarhreppur()
}
\description{
function to load data from vestmannaeyjar, Iceland.
the period here is 24 weeks, therefore IP should be (52/24)
}

