% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logcorr_function.R
\name{logcorr}
\alias{logcorr}
\title{logcorr}
\usage{
logcorr(sim)
}
\arguments{
\item{sim}{The dataframe or list produced by the 'runtsir' function.}
}
\description{
Plot the correlation of the true data against the fitted resimulated data.
}

