% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/maxthreshold_function.R
\name{maxthreshold}
\alias{maxthreshold}
\title{maxthreshold Function
used to optimize the threshold parameter to give the best fit to the data}
\usage{
maxthreshold(data, nsim = 2, IP = 2, method = "deterministic", parms,
  thresholdmin = 2, thresholdmax = 20, printon = FALSE)
}
\arguments{
\item{data}{is the time, cases, births, pop data frame}

\item{nsim}{is the number of simulations to do}

\item{IP}{is the infectious period}

\item{method}{is the simulation method used, i.e. deterministic, negbin, pois}

\item{parms}{is the estimated parameters from estpars or mcmcestpars}

\item{thresholdmin}{is the minimum number of cases to be considered an outbreak}

\item{thresholdmax}{is the max number of cases to be considered an outbreak}

\item{printon}{is a T/F statement to print the progress}
}
\description{
maxthreshold Function
used to optimize the threshold parameter to give the best fit to the data
}

