% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subplots_functions.R
\name{plotbeta}
\alias{plotbeta}
\title{plotbeta}
\usage{
plotbeta(dat)
}
\arguments{
\item{dat}{the list produced from the runtsir, mcmctsir, and simulatetsir function.}
}
\description{
Plots the inferred beta with confidence intervals (when they can be calculated)
}
