% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plotbreaks_function.R
\name{plotbreaks}
\alias{plotbreaks}
\title{plotbreaks function}
\usage{
plotbreaks(data, threshold)
}
\arguments{
\item{data,}{dataframe with the cases vector.}

\item{threshold,}{the epidemic threshold.}
}
\description{
plotbreaks function
}

