% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plotcases_function.R
\name{plotcases}
\alias{plotcases}
\title{plotcases Function
plots just the cases}
\usage{
plotcases(data)
}
\arguments{
\item{data}{is the dataframe with cases}
}
\description{
plotcases Function
plots just the cases
}

