% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plotcomp_function.R
\name{plotcomp}
\alias{plotcomp}
\title{plotcomp}
\usage{
plotcomp(sim, errtype = "95")
}
\arguments{
\item{sim}{is list produced by runtsir or mcmctsir}

\item{errtype}{is the type of error bands to show. Defaults to '95' for 95 percent CI, the other option is 'sd' to standard deviation.}
}
\description{
Plots just the comparison of the forward simulation fit to the data.
}

