% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdata_function.R
\name{plotdata}
\alias{plotdata}
\title{plotdata}
\usage{
plotdata(data)
}
\arguments{
\item{data}{The dataframe with time, cases, births, and pop.}
}
\description{
Plots the cases data as well as birth and population dynamics.
}
