% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subplots_functions.R
\name{plotforward}
\alias{plotforward}
\title{plotforward}
\usage{
plotforward(dat, inverse = F)
}
\arguments{
\item{dat}{the list produced from the runtsir, mcmctsir, and simulatetsir function.}

\item{inverse}{a TRUE or FALSE option to plot the forward simulate negative (TRUE) or positive (FALSE). Defaults to FALSE.}
}
\description{
Plots the forward simulation from the TSIR model
}
