% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subplots_functions.R
\name{plotregression}
\alias{plotregression}
\title{plotregression}
\usage{
plotregression(dat)
}
\arguments{
\item{dat}{the list produced from the runtsir, mcmctsir, and simulatetsir function.}
}
\description{
Plots the cumulative cases - cumulative births data and regression fit
}

