% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subplots_functions.R
\name{plotrho}
\alias{plotrho}
\title{plotrho}
\usage{
plotrho(dat)
}
\arguments{
\item{dat}{the list produced from the runtsir, mcmctsir, and simulatetsir function.}
}
\description{
Plots the inferred reporting rate, rho
}

