% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subplots_functions.R
\name{plotsbar}
\alias{plotsbar}
\title{plotsbar}
\usage{
plotsbar(dat)
}
\arguments{
\item{dat}{the list produced from the runtsir, mcmctsir, and simulatetsir function.}
}
\description{
Plots the profile log likelihood calculation for inferred sbar
}
