% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/residual.births_function.R
\name{residual.births}
\alias{residual.births}
\title{residuals.births Function
computes the residuals for when X is the cumulative births
used internally}
\usage{
residual.births(rho, Yhat, Y)
}
\arguments{
\item{rho}{is the reporting rate to get units correct}

\item{Yhat}{is the fitted regression}

\item{Y}{is the cumulative cases}
}
\description{
residuals.births Function
computes the residuals for when X is the cumulative births
used internally
}

