% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/residual.cases_function.R
\name{residual.cases}
\alias{residual.cases}
\title{residuals.cases Function
computes the residuals for when X is the cumulative cases
used internally}
\usage{
residual.cases(Yhat, Y)
}
\arguments{
\item{Yhat}{is the fitted regression}

\item{Y}{is the cumulative births}
}
\description{
residuals.cases Function
computes the residuals for when X is the cumulative cases
used internally
}

