% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/simulatetsir_function.R
\name{simulatetsir}
\alias{simulatetsir}
\title{simulatetsir function
This function just simulates the forward prediction}
\usage{
simulatetsir(data, nsim = 100, IP = 2, parms, method = "deterministic",
  epidemics = "cont", pred = "forward", threshold = 1, add.noise.sd = 0,
  mul.noise.sd = 0)
}
\arguments{
\item{data}{the data frame containing cases and interpolated births and populations.}

\item{nsim}{the number of simulations to do. Defaults to 100.}

\item{IP}{is the infectious period. Defaults to 2.}

\item{parms}{either the parameters estimated by estpars or mcmcestpars, or a list containing
beta, rho, Z, sbar, alpha, X, Y, Yhat, contact, alphalow, alphahigh, loglik, pop vectors.}

\item{method}{the type of next step prediction used. Options are 'negbin' for negative binomial,
'pois' for poisson distribution, and 'deterministic'. Defaults to 'deterministic'.}

\item{epidemics}{the type of data splitting. Options are 'cont' which doesn't split the data up at all,
and 'break' which breaks the epidemics up if there are a lot of zeros. Defaults to 'cont'.}

\item{pred}{the type of prediction used. Options are 'forward' and 'step-ahead'. Defaults to 'forward'.}

\item{threshold}{the cut off for a new epidemic if epidemics = 'break'. Defaults to 1.}

\item{add.noise.sd}{the sd for additive noise, defaults to zero.}

\item{mul.noise.sd}{the sd for multiplicative noise, defaults to zero.}
}
\description{
simulatetsir function
This function just simulates the forward prediction
}

