% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSIR.R
\name{TSIR}
\alias{TSIR}
\title{Multivariate Time Series function}
\usage{
TSIR(X, y, n_slices = 10, n_lags = 12)
}
\arguments{
\item{X}{Predictors}

\item{y}{Target}
}
\description{
Dimension reduction for multivariate time series using sliced inverse regression.
}
\keyword{TS}
