% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/tsiRdata_function.R
\name{tsiRdata}
\alias{tsiRdata}
\title{tsiRdata function
this function takes in time cases births and pop vectors (of any lengths) and interpolates them using the given infectious period}
\usage{
tsiRdata(time, cases, births, pop, IP = 2)
}
\arguments{
\item{time}{the time vector}

\item{cases}{the cases vector}

\item{births}{the births vector}

\item{pop}{the population vector}

\item{IP}{the infectious period to discretize to. Defaults to 2}
}
\description{
tsiRdata function
this function takes in time cases births and pop vectors (of any lengths) and interpolates them using the given infectious period
}

