\name{Measles incidence data from 20 cities}
\alias{twentymeas}
\docType{data}
\title{
22 year times series for 20 cities, already interpolated}
\description{
twentymeas is a list containing 20 dataframes with cases, births, populations.
}
\usage{data("twentymeas")}


\source{
From Bryan Grenfell}

\examples{
data(twentymeas)
london <- twentymeas[["London"]]}
\keyword{datasets}
