% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vplayout_function.R
\name{vplayout}
\alias{vplayout}
\title{vplayout}
\usage{
vplayout(x, y)
}
\arguments{
\item{x}{is the x location of the plot}

\item{y}{is the y lcoation of the ploy}
}
\description{
the function just breaks up the plot area into a grid. Called internally.
}

