% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/derivative_function.R
\name{derivative}
\alias{derivative}
\title{derivative}
\usage{
derivative(X, Y)
}
\arguments{
\item{X}{The variable to differentiate with respect to.}

\item{Y}{The function / vector to differentiate.}
}
\description{
This function computes an 8 point derivative.
}

