% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/estpars_function.R
\name{estpars}
\alias{estpars}
\title{estpars function}
\usage{
estpars(data, xreg = "cumcases", IP = 2, regtype = "gaussian",
  sigmamax = 3, userYhat = numeric(), fit = "glm", sbar = 0.05,
  fittype = "all", printon = F)
}
\arguments{
\item{data,}{the data frame containing cases and interpolated births and populations.}

\item{xreg,}{the x-axis for the regression. Options are 'cumcases' and 'cumbirths'. Defaults to 'cumcases'.}

\item{IP,}{the infectious period in weeks. Defaults to 2 weeks.}

\item{regtype,}{the type of regression used in susceptible reconstruction.
Options are 'gaussian', 'lm' (linear model), 'spline' (smooth.spline with 2.5 degrees freedom),
'lowess' (with f = 2/3, iter = 1), 'loess' (degree 1), and 'user' which is just a user inputed vector.
Defaults to 'gaussian' and if that fails then defaults to loess.}

\item{sigmamax,}{the inverse kernal width for the gaussian regression. Default is 3.
Smaller, stochastic outbreaks tend to need a lower sigma.}

\item{userYhat,}{the inputed regression vector if regtype='user'. Defaults to NULL.}

\item{fit,}{the fitting method used. Options are 'glm' or 'bayesglm' which is a cheap adaptation
to include some bayesian approaches. For 'bayesglm' we use a gaussian prior with mean 1e-4.}

\item{sbar,}{the mean number of susceptibles. Only used if fittype='less'. Defaults to 0.05*mean(pop).}

\item{fittype,}{the type of fit used. Options are 'all' which fits beta, sbar, and alpha,
'fixalpha', which fixes alpha at 0.97 and estimates beta and sbar, and
'less' which fits only beta and fixes alpha at 0.97.}

\item{printon,}{whether to show diagnostic prints or not, defaults to FALSE.}
}
\description{
This function computes the set up to run the TSIR model, i.e. reconstructes susecptibles and
estimates beta and alpha.
}

