% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagsfilter_function.R
\name{jagsfilter}
\alias{jagsfilter}
\title{jagsfilter}
\usage{
jagsfilter(mcmcresults)
}
\arguments{
\item{mcmcresults}{is the input from the jags model.}
}
\description{
Used internally to filter jags results to give just the inference well use.
}
