% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotbreaks_function.R
\name{plotbreaks}
\alias{plotbreaks}
\title{plotbreaks}
\usage{
plotbreaks(data, threshold)
}
\arguments{
\item{data}{Data frame with the cases vector.}

\item{threshold}{The epidemic threshold, i.e. the number of cases required to spark a new outbreak in the model.}
}
\description{
Plots the cases data with a line whenever the forward simulation is seeded using the real data.
}

