% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcomp_function.R
\name{plotcomp}
\alias{plotcomp}
\title{plotcomp}
\usage{
plotcomp(sim, errtype = "95", max.plot = 10)
}
\arguments{
\item{sim}{is list produced by runtsir or mcmctsir}

\item{errtype}{is the type of error bands to show. Defaults to '95' for 95 percent CI, the other option is 'sd' to standard deviation.}

\item{max.plot}{the number of individual stochastic simulations to plot. Defaults to 10.}
}
\description{
Plots just the comparison of the forward simulation fit to the data.
}
