% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plotres_function.R
\name{plotres}
\alias{plotres}
\title{plotres function}
\usage{
plotres(dat)
}
\arguments{
\item{dat,}{the list produced from the runtsir function}
}
\description{
function to plot diagnostics and results of the runtsir function.
}

