% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotres_function.R
\name{plotres}
\alias{plotres}
\title{plotres}
\usage{
plotres(dat, max.plot = 10)
}
\arguments{
\item{dat}{the list produced from the runtsir, mcmctsir, and simulatetsir function.}

\item{max.plot}{the number of individual stochastic simulations to plot. Defaults to 10.}
}
\description{
Plots diagnostics and results of the runtsir function.
}
