% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual.births_function.R
\name{residual.births}
\alias{residual.births}
\title{residuals.births}
\usage{
residual.births(rho, Yhat, Y)
}
\arguments{
\item{rho}{The reporting rate, used to get units correct.}

\item{Yhat}{The fitted regression line.}

\item{Y}{The cumulative cases.}
}
\description{
computes the residuals for when X is the cumulative births. Used internally.
}
