% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual.cases_function.R
\name{residual.cases}
\alias{residual.cases}
\title{residuals.cases}
\usage{
residual.cases(Yhat, Y)
}
\arguments{
\item{Yhat}{The fitted regression line.}

\item{Y}{The cumulative births.}
}
\description{
Computes the residuals for when X is the cumulative cases. Used internally.
}
