% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/runtsir_function.R
\name{runtsir}
\alias{runtsir}
\title{runtsir function}
\usage{
runtsir(data, xreg = "cumcases", IP = 2, nsim = 100,
  regtype = "gaussian", sigmamax = 3, userYhat = numeric(),
  alpha = 0.97, fittype = "all", fit = "glm", family = "gaussian",
  method = "deterministic", epidemics = "cont", pred = "forward",
  threshold = 1, sbar = 0.05, seasonality = "standard",
  add.noise.sd = 0, mul.noise.sd = 0, printon = F)
}
\arguments{
\item{data,}{the data frame containing cases and interpolated births and populations.}

\item{xreg,}{the x-axis for the regression. Options are 'cumcases' and 'cumbirths'. Defaults to 'cumcases'.}

\item{IP,}{the infectious period in weeks. Defaults to 2 weeks.}

\item{nsim,}{the number of simulations to do. Defaults to 100.}

\item{regtype,}{the type of regression used in susceptible reconstruction.
Options are 'gaussian', 'lm' (linear model), 'spline' (smooth.spline with 2.5 degrees freedom),
'lowess' (with f = 2/3, iter = 1), 'loess' (degree 1), and 'user' which is just a user inputed vector.
Defaults to 'gaussian' and if that fails then defaults to loess.}

\item{sigmamax,}{the inverse kernal width for the gaussian regression. Default is 3.
Smaller, stochastic outbreaks tend to need a lower sigma.}

\item{userYhat,}{the inputed regression vector if regtype='user'. Defaults to NULL.}

\item{fittype,}{the type of fit used. Options are 'all' which fits beta, sbar, and alpha,
'fixalpha', which fixes alpha at 0.97 and estimates beta and sbar, and
'less' which fits only beta and fixes alpha at 0.97.}

\item{fit,}{the fitting method used. Options are 'glm' or 'bayesglm' which is a cheap adaptation
to include some bayesian approaches. For 'bayesglm' we use a gaussian prior with mean 1e-4.}

\item{family,}{the family in the GLM regression, options are poisson and gaussian both with
log link. Default is Poisson.}

\item{method,}{the type of next step prediction used. Options are 'negbin' for negative binomial,
'pois' for poisson distribution, and 'deterministic'. Defaults to 'deterministic'.}

\item{epidemics,}{the type of data splitting. Options are 'cont' which doesn't split the data up at all,
and 'break' which breaks the epidemics up if there are a lot of zeros. Defaults to 'cont'.}

\item{pred,}{the type of prediction used. Options are 'forward' and 'step-ahead'. Defaults to 'forward'.}

\item{threshold,}{the cut off for a new epidemic if epidemics = 'break'. Defaults to 1.}

\item{sbar,}{the mean number of susceptibles. Only used if fittype='less'. Defaults to 0.05*mean(pop).}

\item{seasonality,}{the type of contact to use. Options are standard for 52/IP point contact or schoolterm for just a two point on off contact. Defaults to standard.}

\item{add.noise.sd,}{the sd for additive noise, defaults to zero.}

\item{mul.noise.sd,}{the sd for multiplicative noise, defaults to zero.}

\item{printon,}{whether to show diagnostic prints or not, defaults to FALSE.}
}
\description{
This function runs the TSIR model
}

