% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsiRdata_function.R
\name{tsiRdata}
\alias{tsiRdata}
\title{tsiRdata}
\usage{
tsiRdata(time, cases, births, pop, IP = 2)
}
\arguments{
\item{time}{The time vector.}

\item{cases}{The cases vector.}

\item{births}{The births vector.}

\item{pop}{The population vector.}

\item{IP}{The infectious period (in weeks) to discretize to. Defaults to 2.}
}
\description{
A function to take in time cases births and pop vectors (of any lengths) and interpolate them using the given infectious period.
}

